move slatedb wal stuff under slatedb dir
