+++ slatedb/src/fence.rs
@@ -3,7 +3,7 @@ use crate::error::SlateDBError;
 use crate::manifest::store::{FenceableManifest, StoredManifest};
 use crate::tablestore::TableStore;
 use crate::utils::WatchableOnceCellReader;
-use crate::wal::writer_init::{WalWriterInit, WalWriterInitOptions};
+use crate::wal::slatedb::writer_init::{WalWriterInit, WalWriterInitOptions};
 use crate::wal::{WalIterator, WalWriter, WriterInit};
 use crate::Settings;
 use fail_parallel::{fail_point_send, FailPointTx};
