+++ slatedb/src/wal/slatedb/mod.rs
@@ -1,0 +1,7 @@
+//! SlateDB's native object-store-backed WAL implementation.
+
+pub(crate) mod admin;
+pub(crate) mod gc;
+pub(crate) mod reader;
+pub(crate) mod wal_sst_builder;
+pub(crate) mod writer_init;
