+++ slatedb/src/wal/slatedb/writer_init.rs
